DIRECTOR: Move the saving functionality from archive.cpp to archive-save.cpp
